# Use random to generate keys for the benchmark