In real life the keys would probably come with different periods, than just sequentially

This also prepares us better should we try _other_ LRU cache implementations for a more realistic comparison.