@vagababov 2020-08-17
Actually quite different
```
>benchcmp old.txt new.txt
benchmark                                                     old ns/op     new ns/op     delta
BenchmarkPodRevisionContext/sequential-1-revisions-4          323           334           +3.41%
BenchmarkPodRevisionContext/parallel-1-revisions-4            403           404           +0.25%
BenchmarkPodRevisionContext/sequential-4095-revisions-4       494           493           -0.20%
BenchmarkPodRevisionContext/parallel-4095-revisions-4         562           732           +30.25%
BenchmarkPodRevisionContext/sequential-409600-revisions-4     2181          2094          -3.99%
BenchmarkPodRevisionContext/parallel-409600-revisions-4       2393          2466          +3.05%
```